chore(tests for graphimpl): removed tests for old architecture, and add new ones
